fetching csv in import subscrier action instead of Importer Go func
